User relational model at models.py
