+++ sellotape/sellotape_dj/main_app/models.py
@@ -1,14 +1,32 @@
 from django.db import models
 from django.contrib.auth.models import User
-from .admin import UserProfile
+
+# https://docs.djangoproject.com/en/dev/topics/auth/customizing/#extending-the-existing-user-model
+class Profile(models.Model):
+    # Inherits Attributes from User default instance
+    user = models.OneToOneField(User, on_delete=models.CASCADE)
+    # User avatar, optional: limit the image instance height and width
+    avatar = models.ImageField()
+    country = models.CharField(max_length=100, default='Israel')
+    
+    class City(models.IntegerChoices):
+        tel_aviv = '1', 'Tel Aviv'
+        hamerkaz = '2', 'HaMerkaz'
+        hadarom = '3', 'HaDarom'
+        hatzafon = '4', 'HaTzafon'
+    
+    city = models.IntegerField(
+        choices=City.choices,
+        default=City.tel_aviv,
+    )   
 
 class UserFollowers(models.Model):
     user_followers_id = models.AutoField(primary_key=True)
     user_id = models.ForeignKey('self', on_delete=models.CASCADE)
-    follower_id	= models.ForeignKey(UserProfile, on_delete=models.CASCADE)
+    follower_id	= models.ForeignKey(Profile, on_delete=models.CASCADE)
 
 class Stream(models.Model):
-    author = models.ForeignKey(UserProfile, on_delete=models.CASCADE)
+    author = models.ForeignKey(Profile, on_delete=models.CASCADE)
     link = models.URLField(max_length=500, blank=False)
     airs_on = models.DateField()
     ends_on = models.DateField(null=True)
